I broke something :(
git-svn-id: http://svn.code.sf.net/p/distributedcl/code/trunk@194 cb344d40-e590-48b6-843f-7a98a4fc1ff3

+++ src/libdistributedcl/network/platform.h
@@ -30,7 +30,7 @@ namespace dcl {
 namespace network {
 namespace platform {
 //-----------------------------------------------------------------------------
-typedef uint16_t session_id_t;
+typedef uint32_t session_id_t;
 typedef uint16_t session_sequence_t;
 //-----------------------------------------------------------------------------
 }}} // namespace dcl::network::platform;
